// src/lib/api/schemas/salesOrderStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */

export type SalesOrderStatus = typeof SalesOrderStatus[keyof typeof SalesOrderStatus];


export const SalesOrderStatus = {
  draft: 'draft',
  quote: 'quote',
  confirmed: 'confirmed',
  partially_fulfilled: 'partially_fulfilled',
  fulfilled: 'fulfilled',
  cancelled: 'cancelled',
  closed: 'closed',
} as const;